+++ doc/releaseNotes.md
@@ -445,3 +445,4 @@ This file describes the main feature changes for each readout.exe released versi
 - added support for custom commands to be executed on states transitions.
 - Updated configuration parameters:
   - readout.flushConsumerTimeout: when set, readout waits up to this amount of time that all data pages locked by consumers are released before stopping.
+- Added warning message on buffers low.